Add integrity checks to recovery
Introduce the required integrity checking into the recovery procedure,
returning errors when issues are found. We'll definitely need to improve
the error messaging and add tests for these, but this provides the basic
structure and validations.

Use the 0-indexed ID instead for consistency in referencing the parties
so we don't have to recall if it's +1 or not except for the two places
we do the sharing.

Leave some comments with open questions to send over to Phil